@@ -26,11 +26,12 @@
     "babelify": "^7.3.0",
     "browserify-shim": "^3.8.14",
     "envify": "^4.1.0",
-    "enzyme": "^2.9.1",
+    "enzyme": "^3.1.0",
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-react": "^7.3.0",
     "grunt": "^1.0.1",
+    "grunt-apidoc": "^0.11.0",
     "grunt-browserify": "^5.2.0",
     "grunt-cli": "^1.2.0",
     "grunt-contrib-clean": "^1.1.0",
@@ -56,14 +57,15 @@
     "lodash": "^4.17.4",
     "moment": "^2.18.1",
     "ol": "^4.3.3",
+    "path-to-regexp": "^2.0.0",
     "prop-types": "^15.5.10",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-dropzone": "^4.1.3",
     "react-intl": "^2.4.0",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
-    "react-router-redux": "^4.0.8",
+    "react-router-redux": "^5.0.0-alpha.8",
     "react-select": "^1.0.0-rc.10",
     "react-table": "^6.5.3",
     "react-toastify": "^2.0.0",
@@ -75,7 +77,7 @@
     "redux-thunk": "^2.2.0",
     "tether": "^1.4.0",
     "url-search-params": "^0.10.0",
-    "victory": "^0.22.2"
+    "victory": "^0.24.0"
   },
   "browserify": {
     "transform": [
