@@ -1,8 +1,18 @@
 import * as React from 'react';
 import * as ReactRedux from 'react-redux';
+import { FormattedTime } from 'react-intl';
 import { bindActionCreators } from 'redux';
+import {
+  Card, CardBlock, CardTitle, Row, Col,
+  ButtonToolbar, Button, ButtonGroup, Label, Input
+} from 'reactstrap';
+
+import Placeholder from './placeholder';
+
+import moment from 'moment';
 
 import { fetchResources } from '../../ducks/data/resources';
+
 /**
  * Browse and manage resources
  *
@@ -19,13 +29,37 @@ class ResourceExplorer extends React.Component {
     const { resources } = this.props;
     return (
       <div className="animated fadeIn">
-        <ul>
-          {
-            resources.map(resource => (
-              <li key={resource.id}>{resource.name}: {resource.description} ({resource.format})</li>
-            ))
-          }
-        </ul>
+        <Row>
+          <Col className="col-12">
+            <Card>
+              <CardBlock className="card-body">
+                <Row className="mb-2">
+                  <Col >
+                    <div className="small text-muted">Last Update: <FormattedTime value={moment().toDate()} day='numeric' month='numeric' year='numeric' /></div>
+                  </Col>
+                </Row>
+                <Row style={{ height: 200 }} className="mb-2">
+                  <Col>
+                    <Placeholder label="Filter" iconClass="fa fa-filter" />
+                  </Col>
+                </Row>
+                <Row style={{ height: 400 }} className="mb-2">
+                  <Col>
+                    <Placeholder label="Resources" iconClass="fa fa-table" />
+                  </Col>
+                  <Col>
+                    <Placeholder label="Map" iconClass="fa fa-map-o" />
+                  </Col>
+                </Row>
+                <Row style={{ height: 400 }} className="mb-2">
+                  <Col>
+                    <Placeholder label="Details" iconClass="fa fa-database" />
+                  </Col>
+                </Row>
+              </CardBlock>
+            </Card>
+          </Col>
+        </Row >
       </div>
     );
   }
